requirements.txt
deepgram-sdk
Flask
Flask-Cors
youtube_dl